backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.20.0
pydantic==1.10.13
yfinance==0.2.28
openai==0.28.1
python-dotenv==0.21.1
